backend/requirements.txt
# Core Flask
Flask==3.0.0
Flask-SocketIO==5.3.6
python-socketio==5.10.0
eventlet==0.34.0

# Database
Flask-PyMongo==2.3.0
pymongo==4.6.0

# Authentication
Flask-JWT-Extended==4.6.0
bcrypt==4.1.0
PyJWT==2.8.0

# HTTP/API
requests==2.31.0
httpx==0.25.0

# Validation
marshmallow==3.20.0
email-validator==2.1.0

# File handling
Pillow==10.1.0

# CORS
Flask-Cors==4.0.0

# Rate limiting
Flask-Limiter==3.5.0

# Environment
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2

# Production
gunicorn==21.2.0
gevent==23.9.1
gevent-websocket==0.10.1
